# specs/001-nightlife-directory/contracts/api.yaml
...
openapi: 3.0.3
info:
  title: Nightnice API
  description: |
    REST API for Thailand Nightlife Directory.
    - Public endpoints: No authentication required
    - Admin endpoints: JWT Bearer token required
  version: 1.0.0
  contact:
    name: Nightnice Team
servers:
  - url: https://api.nightnice.com
    description: Production
  - url: https://api-staging.nightnice.com
    description: Staging
  - url: http://localhost:5000
    description: Local Development
...
paths:
...
  /api/stores:
    get:
      tags: [Stores]
      summary: List stores with filters
      description: Get paginated list of active stores with optional filters
      parameters:
        - name: province
          in: query
          schema:
            type: string
          description: Filter by province slug
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category slug
        - name: search
          in: query
          schema:
            type: string
            maxLength: 100
          description: Search term (name, description)
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          description: Items per page
        - name: featured
          in: query
          schema:
            type: boolean
          description: Filter featured stores only
        - name: sort
          in: query
          schema:
            type: string
            enum: [name, created_at, -name, -created_at]
            default: -created_at
          description: Sort field (prefix - for descending)
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/stores/{slug}:
    get:
      tags: [Stores]
      summary: Get store by slug
      description: Get detailed store information by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/provinces/{slug}:
    get:
      tags: [Provinces]
      summary: Get province by slug
      description: Get province details with store count by category
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProvinceDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/categories/{slug}:
    get:
      tags: [Categories]
      summary: Get category by slug
      description: Get category details with store count by province
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/ads:
    get:
      tags: [Advertisements]
      summary: Get active advertisements
      description: Get ads targeted for the current page context
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [banner, sponsored, featured]
          description: Filter by ad type
        - name: province
          in: query
          schema:
            type: string
          description: Province slug for targeting
        - name: category
          in: query
          schema:
            type: string
          description: Category slug for targeting
        - name: page_context
          in: query
          schema:
            type: string
            maxLength: 100
          description: Current page context for tracking
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdListResponse'
...
  /api/admin/stores:
    get:
      tags: [Admin Stores]
      summary: List all stores (Admin)
      description: Get all stores including inactive ones
      security:
        - BearerAuth: []
      parameters:
        - name: province
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: is_active
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminStoreListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Admin Stores]
      summary: Create store
      description: Create a new store
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreCreateRequest'
      responses:
        '201':
          description: Store created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminStoreResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/admin/stores/{id}:
    get:
      tags: [Admin Stores]
      summary: Get store by ID (Admin)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminStoreResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Admin Stores]
      summary: Update store
      description: Update an existing store
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreUpdateRequest'
      responses:
        '200':
          description: Store updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminStoreResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Admin Stores]
      summary: Delete store
      description: Permanently delete a store
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Store deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/admin/stores/{id}/images:
    post:
      tags: [Admin Stores]
      summary: Upload store image
      description: Upload an image to store gallery
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                type:
                  type: string
                  enum: [logo, banner, gallery]
              required:
                - file
                - type
      responses:
        '201':
          description: Image uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageUploadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/admin/stores/{id}/images/{imageId}:
    delete:
      tags: [Admin Stores]
      summary: Delete store image
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: imageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Image deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================
  # ADMIN PROVINCE MANAGEMENT
  # ============================================
  /api/admin/provinces:
    get:
      tags: [Admin Provinces]
      summary: List all provinces (Admin)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminProvinceListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/admin/provinces/{id}:
    put:
      tags: [Admin Provinces]
      summary: Update province
      description: Update province SEO description and sort order
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProvinceUpdateRequest'
      responses:
        '200':
          description: Province updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminProvinceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================
  # ADMIN CATEGORY MANAGEMENT
  # ============================================
  /api/admin/categories:
    get:
      tags: [Admin Categories]
      summary: List all categories (Admin)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCategoryListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Admin Categories]
      summary: Create category
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreateRequest'
      responses:
        '201':
          description: Category created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCategoryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - invalid or missing token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    StoreListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/StoreListItem'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
      required:
        - data
        - meta
...
    CategoryDetailResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        province_counts:
          type: array
          items:
            type: object
            properties:
              province:
                $ref: '#/components/schemas/ProvinceRef'
              count:
                type: integer
      required:
        - id
        - name
        - slug
        - province_counts
...